Rename servo to serial
